ai_generator.py
# ai_generator.py
from google import genai
import config
import re
import requests
from bs4 import BeautifulSoup

client = genai.Client(api_key=config.GEMINI_KEY)
MODEL_ID = "gemini-2.5-flash"

# Словарь с разными личностями (шаблонами)
PROMPTS = {
    "uz": """Ты — креативный редактор Telegram-канала о модах для Minecraft.
Я передам тебе текст. Вычлени главное и напиши пост. Уложись в 800 символов.
Пиши ТОЛЬКО на узбекском латинице. Если не смог найти версию напиши "1.21+".
Используй тег <blockquote expandable> для основного блока.

Формат:
📦 <b>[Название мода]</b>

<blockquote expandable><b>Bu nima?</b>
[Описание]

<b>Asosiy xususiyatlar:</b>
• [Фишка 1]
• [Фишка 2]

🎮 Versiya: [Версия]</blockquote>

<blockquote>💖 - zo`r
💔 - Unchamas</blockquote>

#[Хэштег1] #[Хэштег2]
""",

    "ru": """Ты — креативный редактор Telegram-канала о модах для Minecraft.
Я передам тебе текст. Вычлени главное и напиши пост в драйвовом и веселом стиле. Уложись в 800 символов.
Пиши ТОЛЬКО на русском языке.
Используй тег <blockquote expandable> для основного блока.

Формат:
📦 <b>[Название мода]</b>

<blockquote expandable><b>Что это такое?</b>
[Описание]

<b>Главные фишки:</b>
• [Фишка 1]
• [Фишка 2]

🎮 Версия: [Версия]</blockquote>

<blockquote>💖 - Имба
💔 - Не оч</blockquote>

#[Хэштег1] #[Хэштег2]
""",

    "en": """You are a creative editor for a Minecraft mods Telegram channel.
Extract the main points and write an engaging post. Keep it under 800 characters.
Write ONLY in English in an exciting tone.
Use the <blockquote expandable> tag for the main body.

Format:
📦 <b>[Mod Name]</b>

<blockquote expandable><b>What is it?</b>
[Description]

<b>Key Features:</b>
• [Feature 1]
• [Feature 2]

🎮 Version: [Version]</blockquote>

<blockquote>💖 - Awesome
💔 - Not great</blockquote>

#[Hashtag1] #[Hashtag2]
"""
}

def extract_url(text):
    urls = re.findall(r'(https?://[^\s]+)', text)
    return urls[0] if urls else None

def fetch_page_content(url):
    try:
        headers = {'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64)'}
        response = requests.get(url, headers=headers, timeout=10)
        soup = BeautifulSoup(response.content, 'html.parser')
        text = soup.get_text(separator=' ', strip=True)
        return text[:5000] 
    except Exception as e:
        print(f"⚠️ Не смог прочитать сайт {url}: {e}")
        return ""

# Теперь функция принимает параметр persona
def generate_post(user_input, persona="uz"):
    url = extract_url(user_input)
    site_context = ""
    
    if url:
        print(f"🔗 Найдена ссылка, читаю сайт: {url}")
        page_text = fetch_page_content(url)

    # Достаем нужный шаблон по ключу (uz, ru, en)
    selected_prompt = PROMPTS.get(persona, PROMPTS["uz"])
    
    prompt = f"{selected_prompt}\n\nСырая информация от пользователя:\n{user_input}{site_context}"
    response = client.models.generate_content(model=MODEL_ID, contents=prompt)
    
    return response.text.strip()